luxeplan/backend/app/services/depth.py
"""
Depth Estimation Service
Monocular depth estimation using MiDaS or DPT for occlusion logic.
"""

from PIL import Image


class DepthEstimationService:
    def __init__(self):
        # In production: load MiDaS or DPT model
        # self.model = DPTForDepthEstimation.from_pretrained("Intel/dpt-large")
        # self.processor = DPTImageProcessor.from_pretrained("Intel/dpt-large")
        self._model_loaded = False

    async def estimate(self, image: Image.Image, image_id: str) -> str:
        """
        Run monocular depth estimation.
        Returns URL to the depth map image.

        Production implementation:
        1. Preprocess image through DPT processor
        2. Run inference to get depth prediction
        3. Normalize depth values to 0-255 range
        4. Save as grayscale PNG
        5. Upload to Supabase storage
        6. Return public URL

        The depth map is used by:
        - PlacementEngine for z-ordering
        - RenderingEngine for occlusion masking
        - Before/after compositing
        """
        # Development stub: return placeholder URL
        return f"/api/depth/{image_id}/depth_map.png"
