luxeplan/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
pillow==11.1.0
numpy==2.2.1
pydantic==2.10.4
httpx==0.28.1
python-dotenv==1.0.1
supabase==2.11.0
google-generativeai==0.8.4
rembg==2.0.62
onnxruntime==1.20.1
transformers==4.47.1
torch==2.5.1
torchvision==0.20.1
opencv-python-headless==4.10.0.84
